ESC-688 move separate builder classes into a dedicated package
